fastapi/requirements.txt
fastapi==0.120.1
uvicorn==0.34.0
python-multipart==0.0.20
yt-dlp==2025.12.08
openai==1.58.1
python-dotenv==1.0.1
weasyprint==63.1
markdown==3.7
aiofiles==24.1.0
pydantic==2.10.4
celery==5.4.0
redis==7.2
